디렉토리 수정및 불필요한 파일 제거
