--- API-Server/api/__pycache__/__init__.py
@@ -1,3 +1,0 @@
-from .app import application
-from .db import db as blockchain_db
-#from .db import redis_db
