--- API-Server/api/util.py
@@ -1,10 +1,0 @@
-ANS_ID = "c56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b"
-ANC_ID = "602c79718b16e442de58778e148d0b1084e3b2dffd5de6b7b16cee7969282de7"
-
-
-PRIVATENET_SEEDS = [
-    "http://localhost:30333",
-    "http://localhost:30334",
-    "http://localhost:30335",
-    "http://localhost:30336",
-]
